requirements.txt
asgiref==3.11.1
boto3==1.34.69
botocore==1.34.162
crispy-bootstrap5==2024.2
Django==5.0.4
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-crispy-forms==2.1
django-environ==0.11.2
django-js-asset==3.1.2
django-meta==2.4.1
django-robots==6.1
django-sitemap-generate==0.7.0
django-storages==1.14.2
gunicorn==21.2.0
jmespath==1.1.0
packaging==26.0
pillow==12.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.6.0
